# target/classes/srs-openapi.yaml
openapi: "3.0.3"
info:
  title: "srs API"
  description: "srs API"
  version: "1.0.0"
servers:
  - url: "https://srs"
paths:
...
  /applications/add:
    post:
      summary: "POST applications/add"
      operationId: "addApplication"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationDTO"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OperationStatus"
  /applications/delete/{id}:
    delete:
      summary: "DELETE applications/delete/{id}"
      operationId: "deleteApplication"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int32"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OperationStatus"
  /applications/filter:
    get:
      summary: "GET applications/filter"
      operationId: "getApplicationsByFilter"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationFilter"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/ApplicationDTO"
  /applications/update:
    put:
      summary: "PUT applications/update"
      operationId: "updateApplication"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationDTO"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OperationStatus"
...
  /applicationTypes/add:
    post:
      summary: "POST applicationTypes/add"
      operationId: "addApplicationType"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationTypeDTO"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OperationStatus"
  /applicationTypes/delete/{id}:
    delete:
      summary: "DELETE applicationTypes/delete/{id}"
      operationId: "deleteApplicationType"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int32"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OperationStatus"
  /applicationTypes/filter:
    get:
      summary: "GET applicationTypes/filter"
      operationId: "getApplicationTypesByFilter"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationTypeFilter"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/ApplicationTypeDTO"
  /applicationTypes/update:
    put:
      summary: "PUT applicationTypes/update"
      operationId: "updateApplicationType"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationTypeDTO"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OperationStatus"
...
  /auth/loginUser:
    post:
      summary: "POST auth/loginUser"
      operationId: "loginUser"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Credentials"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AuthSession"
...
  /auth/refreshToken:
    post:
      summary: "POST auth/refreshToken"
      operationId: "refreshToken"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshToken"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AuthSession"
  /auth/sign-up:
    post:
      summary: "POST auth/sign-up"
      operationId: "registration"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegistrationDTO"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OperationStatus"
...
  /categories/add:
    post:
      summary: "POST categories/add"
      operationId: "addCategory"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryDTO"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OperationStatus"
  /categories/delete/{id}:
    delete:
      summary: "DELETE categories/delete/{id}"
      operationId: "deleteCategory"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int32"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OperationStatus"
  /categories/filter:
    get:
      summary: "GET categories/filter"
      operationId: "getCategoriesByFilter"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryFilter"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/CategoryDTO"
  /categories/update:
    put:
      summary: "PUT categories/update"
      operationId: "updateCategory"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryDTO"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OperationStatus"
components:
  schemas:
...
    OperationStatus:
      type: "object"
      properties:
        targetClassName:
          type: "string"
        recordId:
          type: "integer"
          format: "int32"
        operationName:
          type: "string"
        success:
          type: "boolean"
        msgList:
          type: "array"
          items:
            type: "string"
...
    ApplicationDTO:
      type: "object"
      properties:
        applicationId:
          type: "integer"
          format: "int32"
        prisonerId:
          type: "integer"
          format: "int32"
        additionalInformation:
          type: "string"
        typeOfApplication:
          type: "integer"
          format: "int32"
        userId:
          type: "integer"
          format: "int32"
        applyDate:
          type: "string"
          format: "date"
        reviewDate:
          type: "string"
          format: "date"
    ApplicationFilter:
      type: "object"
      properties:
        applicationIdList:
          type: "array"
          items:
            type: "integer"
            format: "int32"
    ApplicationTypeDTO:
      type: "object"
      properties:
        applicationTypeId:
          type: "integer"
          format: "int32"
        description:
          type: "string"
        name:
          type: "string"
    ApplicationTypeFilter:
      type: "object"
      properties: { }
    GrantedAuthority:
      type: "object"
      properties: { }
    CurrentUser:
      type: "object"
      properties:
        jwtToken:
          type: "string"
        id:
          type: "integer"
          format: "int32"
        firstName:
          type: "string"
        lastName:
          type: "string"
        email:
          type: "string"
        login:
          type: "string"
        password:
          type: "string"
        enabled:
          type: "boolean"
        locked:
          type: "boolean"
        accountExpired:
          type: "string"
          format: "date"
        passwordExpired:
          type: "string"
          format: "date"
        lastLoginTime:
          type: "string"
          format: "date-time"
        userType:
          type: "integer"
          format: "int32"
        grantedAuthorities:
          type: "array"
          items:
            $ref: "#/components/schemas/GrantedAuthority"
    Credentials:
      type: "object"
      properties:
        userName:
          type: "string"
        password:
          type: "string"
    AuthSession:
      type: "object"
      properties:
        accessToken:
          type: "string"
        expiresIn:
          type: "integer"
          format: "int32"
        refreshToken:
          type: "string"
        refreshExpiresIn:
          type: "integer"
          format: "int32"
        currentUser:
          $ref: "#/components/schemas/CurrentUser"
...
    RefreshToken:
      type: "object"
      properties:
        refreshToken:
          type: "string"
    RegistrationDTO:
      type: "object"
      properties:
        userId:
          type: "integer"
          format: "int32"
        locationId:
          type: "integer"
          format: "int32"
        firstname:
          type: "string"
        lastname:
          type: "string"
        email:
          type: "string"
        login:
          type: "string"
        password:
          type: "string"
        userType:
          type: "integer"
          format: "int32"
    CategoryDTO:
      type: "object"
      properties:
        categoryId:
          type: "integer"
          format: "int32"
        code:
          type: "string"
        description:
          type: "string"
        helpDescription:
          type: "string"
    CategoryFilter:
      type: "object"
      properties: { }